project-management-backend/requirements.txt
# Flask Core
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Database - MySQL
pymysql==1.1.0
cryptography==41.0.7

# Validation
marshmallow==3.20.1
python-dotenv==1.0.0

# Security
bcrypt==4.1.2

# Utils
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-flask==1.3.0
black==23.12.1
flake8==7.0.0
